feat: Configurable font size/scale (#72)
Adds a field on the Customize Lamp page to allow users to scale the font
size. The size gets saved to file with the other options and gets
loaded/applied on application start.
This should also reset to the original scale when the reset button is
pressed.

I think the ideal solution would be to change the actual font-size
instead of changing the font scale (scaling can make things blurry), but
I was not sure how to accomplish that.