feat: Configurable alternate button color (#80)
Added a new configurable color `Colour_ButtonAlt`, and used that color
for the "Enabled" button on the mods table.

Having a different color for the enabled/disabled button state can be
useful for quickly determining which mods are and are not enabled at a
glance.

I did not put much consideration/thought into the default color that I
used (#179642FF). I would recommend taking a look and seeing if you
would prefer using a different color as the default.