feat: Add configuration for disabling automatic update check and add a a button for manual checks (#82)
Added an option to disable the automatic check for updates when the
application starts. This also add and item to the menu to manually check
for updates.

By default, the stored value in the configuration file would be a 1 or
0, but the checks would also support "false" as a valid value to turn it
off. If the value in the configuration is not 0 or "false", then we
assume that it is set to true, and do the update check.

I wasn't sure if this is something you would want to include, but
figured I would make a pull request in case you don't mind having it.
For some users (such as myself), this removes (what I would consider)
unnecessary network traffic. I haven't tested this, but it could also
technically provide a slight improvement in the startup time when
disabled, as this check was done prior to displaying the UI.