fix: a crash that can happen when deleting mods (#85)
This should fix a crash that could happen when deleting mods, which
seems to have been introduced with the disabled button change in
65c0c4241a4906884b1fb1265b7bebf0d96da6a8.

I did not fully explore the cause, but it seemed like the crash would
occur when you had at least 1 mod enabled and deleted another mod. I
believe this was because we begin a disabled section (`BeginDisabled`),
but we do not properly close out that section when a mod is deleted.