feat: Highlight mods on hover (#83)
This change attempts to highlight the mod row when you hover over it.
Right now, it uses `Colour_SearchHighlight` for the highlight color.

There are some issues with how this currently works, but I have not
figured out a way to make it work better. Right now, I have it checking
for if the user is hovering over each item in the row individually, and
applying the highlighting to the whole row if they are. This adds some
messiness to the code that I am not a fan of (I feel like there *should*
be some way to check if the user is hovering over the row or any
children in a single check, rather than having to check after each
item). This also means that the highlight effect gets dropped as you
move between elements (for example, if you have a short mod name, the
highlight goes away as you move between the mod name and the mod type).

I do not attempt to apply the highlight when hovering over the mod type,
as that gets interrupted by the type selection menu anyway.

This should also apply the hover for any additional buttons added
through the `ExtraOptions`, but this has not been tested as I am not
really sure how to add those.