feat: Add context menu for moving mods to top or bottom (#93)
This adds a context menu on the mod name with options to move the mod to
the top or bottom of the load order.

This should help with moving mods around on larger mod lists.

I also modified how the drag and drop moves mods to use the new
`moveModTo` function I added. I think this simplifies the logic visually
on top of using the other existing functions for moving mods instead of
relying on something custom.