feat: Add quit button to the menu (#99)
Adds a quit button to the main menu.

This option may feel nicer to some users over requiring the use of the
close button on the window.

I was hoping for a somewhat 'cleaner' way of doing this (such as by
manually triggering an `SDL_QUIT` event from the menu), but I have not
yet figured out if that is possible. If you have a better way of
handling this, feel free to close this PR or let me know. Otherwise, I
believe this should work fine.