feat: Add mod separators (#94)
The primary goal of this PR is to add the ability to add mod separators.

I intend to accomplish this by adding a "mod type" that is mean for mod
separators. This seemed like the most straightforward way to get it
added without needing to potentially touch a bunch of other systems.

## Features

Features that should/need to be done:
- [x] Add mod type for mod separators
- [x] Right click menu with option to add mod separators
- [x] Add separators with default text and mod type
- [x] Ignore separators for some actions (deployment, deleting files,
etc)
- [x] Allow renaming separators
- [x] Fix the mod separator menu overriding the column header context
menu.

Some optional features if I can figure out how to do them and get around
to it:
- [x] Add the separator at (or at least near) where the user has
right-clicked to add it.
- [x] Allow "collapsing" mods below a separator, essentially making them
a grouping of mods.
- The way I would plan to set this up, if possible, is to have it
effectively hide any rows below it until we reach the end of the table
or another mod separator.
- [x] Find a way to allow right-clicking below the table (when there are
few or no mods)
- [x] Possibly tweaking other columns in the mod table for separators,
such as:
- Removing the enable/disable button (and/or possibly replacing it with
a collapse/expand if that feature gets implemented)
	- Hiding the mod type selector
	- Other special handling


## Notes

I have made a number of changes that may not be desirable, so it would
be good to have things looked over to make sure I haven't done anything
overly bad/dumb.

While I was trying to make these changes, I ran into some *issues*
referencing mod types, so I have ended up touching more things than I
would have liked. These changes largely revolve around how mod types are
being handled, referenced, and saved.

Here are some of the additional changes that were made:
- I attempted to create a new variable that would contain the
combination of ModType identifier (from the enum) and the display value
for it to help keep those together.
- My intention was to make it somewhat easier to add new values (you
only edit effectively 1 location), as well as allow re-ordering the mod
types as they are displayed in the mod type selection in the table,
without breaking anyone's mod setup.
- I changed how we get the mod type in the table (as well as for saving)
to actually use the enum value instead of relying on the enum
coincidentally matching the values from the for loop displaying the
options in the mod table.
- Tweaked the `registerArchive` function to have an optional parameter
that allows specifying a mod type to use.
- [possibly something else]


## Potential conflicts

Since this ended up touching some of the files that are likely being
gutted in #88, this may need to wait until those changes are done (or,
if it makes more sense, to be adapted into #88 prior to that getting
merged). This also relies on a right-click menu, so I would probably
want to see #93 merged first if it is good to go, and then the option
this provides can be tweaked to be included into that menu.