feat: Add warning notifications when game paths are not set (#124)
This adds a warning notification when a user goes to deploy mods for a
game, and the game paths are not set.

This warning will only start to show when the user has clicked the
initial "Deploy" button from the mod list screen (ie, it will show when
the user reaches the screen prompting to either start the deployment or
go back). This notification is configured to only display once per
session per game (so if a user clears it, it will not show up again
until they restart Lampray).

To help enable this, this change adds a function to `gameControl` and
each individual game called `installPathSet` which is meant to return
true if all necessary paths are set for a game, and return false if any
are not set.