feat: Re-enable rar support (#144)
This re-enables attempted support for .rar files. It will attempt to
decompress with Rar5 first, and fallback to the older Rar method if that
fails.

If both methods fail, the user will get an error notification with the
error messages from both attempts.

I tested with the main .rar files from
<https://github.com/ssokolow/rar-test-files>. This successfully
extracted the `testfile.rar5.rar` file, but still fails on the
`testfile.rar3.rar` file, so support will probably still be spotty.