feat: changed directory for where data is stored to be fixed (+ some code fixes) (#164)
This initially was supposed to be just be changing where the directory
for data is stored to be fixed, but "Libary" and a header with a space
in it inflicted psychic damage on me.

I'm now working on fixing a crash when trying to deploy my mods in
Baldur's Gate 3. It has something to do with 7zip now. Earlier, it was a
throw from the missing overlay directory for the app data stuff (see
b5c862e for the fix).